Struct google-dfareporting2::AdListMethodBuilder

pub struct AdListMethodBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

Retrieves a list of ads, possibly filtered.

A builder for the list method supported by a ad resource. It is not used directly, but through a AdMethodsBuilder.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.ads().list("profileId")
             .add_type("diam")
             .ssl_required(true)
             .ssl_compliant(true)
             .sort_order("amet.")
             .sort_field("ipsum")
             .add_size_ids("ut")
             .search_string("dolor")
             .add_remarketing_list_ids("sea")
             .add_placement_ids("ut")
             .page_token("eirmod")
             .overridden_event_tag_id("sanctus")
             .max_results(-27)
             .add_landing_page_ids("dolor")
             .add_ids("et")
             .dynamic_click_tracker(false)
             .creative_type("vero")
             .add_creative_optimization_configuration_ids("ut")
             .add_creative_ids("sed")
             .compatibility("et")
             .add_campaign_ids("ipsum")
             .add_audience_segment_ids("justo")
             .archived(true)
             .advertiser_id("vero")
             .active(false)
             .doit();

Methods

impl<'a, C, NC, A> AdListMethodBuilder<'a, C, NC, A> where NC: NetworkConnector, C: BorrowMut<Client<NC>>, A: GetToken

fn doit(self) -> Result<(Response, AdsListResponse)>

Perform the operation you have build so far.

fn profile_id(self, new_value: &str) -> AdListMethodBuilder<'a, C, NC, A>

Sets the profile id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

User profile ID associated with this request.

fn add_type(self, new_value: &str) -> AdListMethodBuilder<'a, C, NC, A>

Append the given value to the type query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

Select only ads with these types.

fn ssl_required(self, new_value: bool) -> AdListMethodBuilder<'a, C, NC, A>

Sets the ssl required query property to the given value.

Select only ads that require SSL.

fn ssl_compliant(self, new_value: bool) -> AdListMethodBuilder<'a, C, NC, A>

Sets the ssl compliant query property to the given value.

Select only ads that are SSL-compliant.

fn sort_order(self, new_value: &str) -> AdListMethodBuilder<'a, C, NC, A>

Sets the sort order query property to the given value.

Order of sorted results, default is ASCENDING.

fn sort_field(self, new_value: &str) -> AdListMethodBuilder<'a, C, NC, A>

Sets the sort field query property to the given value.

Field by which to sort the list.

fn add_size_ids(self, new_value: &str) -> AdListMethodBuilder<'a, C, NC, A>

Append the given value to the size ids query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

Select only ads with these size IDs.

fn search_string(self, new_value: &str) -> AdListMethodBuilder<'a, C, NC, A>

Sets the search string query property to the given value.

Allows searching for objects by name or ID. Wildcards (*) are allowed. For example, "ad*2015" will return objects with names like "ad June 2015", "ad April 2015" or simply "ad 2015". Most of the searches also add wildcards implicitly at the start and the end of the search string. For example, a search string of "ad" will match objects with name "my ad", "ad 2015" or simply "ad".

fn add_remarketing_list_ids(self, new_value: &str) -> AdListMethodBuilder<'a, C, NC, A>

Append the given value to the remarketing list ids query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

Select only ads whose list targeting expression use these remarketing list IDs.

fn add_placement_ids(self, new_value: &str) -> AdListMethodBuilder<'a, C, NC, A>

Append the given value to the placement ids query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

Select only ads with these placement IDs assigned.

fn page_token(self, new_value: &str) -> AdListMethodBuilder<'a, C, NC, A>

Sets the page token query property to the given value.

Value of the nextPageToken from the previous result page.

fn overridden_event_tag_id(self, new_value: &str) -> AdListMethodBuilder<'a, C, NC, A>

Sets the overridden event tag id query property to the given value.

Select only ads with this event tag override ID.

fn max_results(self, new_value: i32) -> AdListMethodBuilder<'a, C, NC, A>

Sets the max results query property to the given value.

Maximum number of results to return.

fn add_landing_page_ids(self, new_value: &str) -> AdListMethodBuilder<'a, C, NC, A>

Append the given value to the landing page ids query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

Select only ads with these landing page IDs.

fn add_ids(self, new_value: &str) -> AdListMethodBuilder<'a, C, NC, A>

Append the given value to the ids query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

Select only ads with these IDs.

fn dynamic_click_tracker(self, new_value: bool) -> AdListMethodBuilder<'a, C, NC, A>

Sets the dynamic click tracker query property to the given value.

Select only dynamic click trackers. Applicable when type is AD_SERVING_CLICK_TRACKER. If true, select dynamic click trackers. If false, select static click trackers. Leave unset to select both.

fn creative_type(self, new_value: &str) -> AdListMethodBuilder<'a, C, NC, A>

Sets the creative type query property to the given value.

Select only ads with the specified creativeType.

fn add_creative_optimization_configuration_ids(self, new_value: &str) -> AdListMethodBuilder<'a, C, NC, A>

Append the given value to the creative optimization configuration ids query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

Select only ads with these creative optimization configuration IDs.

fn add_creative_ids(self, new_value: &str) -> AdListMethodBuilder<'a, C, NC, A>

Append the given value to the creative ids query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

Select only ads with these creative IDs assigned.

fn compatibility(self, new_value: &str) -> AdListMethodBuilder<'a, C, NC, A>

Sets the compatibility query property to the given value.

Select default ads with the specified compatibility. Applicable when type is AD_SERVING_DEFAULT_AD. WEB and WEB_INTERSTITIAL refer to rendering either on desktop or on mobile devices for regular or interstitial ads, respectively. APP and APP_INTERSTITIAL are for rendering in mobile apps. IN_STREAM_VIDEO refers to rendering an in-stream video ads developed with the VAST standard.

fn add_campaign_ids(self, new_value: &str) -> AdListMethodBuilder<'a, C, NC, A>

Append the given value to the campaign ids query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

Select only ads with these campaign IDs.

fn add_audience_segment_ids(self, new_value: &str) -> AdListMethodBuilder<'a, C, NC, A>

Append the given value to the audience segment ids query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

Select only ads with these audience segment IDs.

fn archived(self, new_value: bool) -> AdListMethodBuilder<'a, C, NC, A>

Sets the archived query property to the given value.

Select only archived ads.

fn advertiser_id(self, new_value: &str) -> AdListMethodBuilder<'a, C, NC, A>

Sets the advertiser id query property to the given value.

Select only ads with this advertiser ID.

fn active(self, new_value: bool) -> AdListMethodBuilder<'a, C, NC, A>

Sets the active query property to the given value.

Select only active ads.

fn delegate(self, new_value: &'a mut Delegate) -> AdListMethodBuilder<'a, C, NC, A>

Sets the delegate property to the given value.

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

fn param<T>(self, name: T, value: T) -> AdListMethodBuilder<'a, C, NC, A> where T: Str

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> AdListMethodBuilder<'a, C, NC, A> where T: Str

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead of relying on the automated algorithm which simply prefers read-only scopes over those who are not.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.