Struct google-dfareporting2::DirectorySite

pub struct DirectorySite {
    pub kind: Option<String>,
    pub country_id: Option<String>,
    pub description: Option<String>,
    pub settings: Option<DirectorySiteSettings>,
    pub currency_id: Option<String>,
    pub inpage_tag_formats: Vec<String>,
    pub url: Option<String>,
    pub contact_assignments: Vec<DirectorySiteContactAssignment>,
    pub parent_id: Option<String>,
    pub interstitial_tag_formats: Vec<String>,
    pub id_dimension_value: Option<DimensionValue>,
    pub active: Option<bool>,
    pub id: Option<String>,
    pub name: Option<String>,
}

DirectorySites contains properties of a website from the Site Directory. Sites need to be added to an account via the Sites resource before they can be assigned to a placement.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get directory sites (response)

list directory sites (none)

Fields

kind

Identifies what kind of resource this is. Value: the fixed string "dfareporting#directorySite".

country_id

Country ID of this directory site.

description

Description of this directory site.

settings

Directory site settings.

currency_id

Currency ID of this directory site. Possible values are: - "1" for USD - "2" for GBP - "3" for ESP - "4" for SEK - "5" for CAD - "6" for JPY - "7" for DEM - "8" for AUD - "9" for FRF - "10" for ITL - "11" for DKK - "12" for NOK - "13" for FIM - "14" for ZAR - "15" for IEP - "16" for NLG - "17" for EUR - "18" for KRW - "19" for TWD - "20" for SGD - "21" for CNY - "22" for HKD - "23" for NZD - "24" for MYR - "25" for BRL - "26" for PTE - "27" for MXP - "28" for CLP - "29" for TRY - "30" for ARS - "31" for PEN - "32" for ILS - "33" for CHF - "34" for VEF - "35" for COP - "36" for GTQ

inpage_tag_formats

Tag types for regular placements.

Acceptable values are: - "STANDARD" - "IFRAME_JAVASCRIPT_INPAGE" - "INTERNAL_REDIRECT_INPAGE" - "JAVASCRIPT_INPAGE"

url

URL of this directory site.

contact_assignments

Directory site contacts.

parent_id

Parent directory site ID.

interstitial_tag_formats

Tag types for interstitial placements.

Acceptable values are: - "IFRAME_JAVASCRIPT_INTERSTITIAL" - "INTERNAL_REDIRECT_INTERSTITIAL" - "JAVASCRIPT_INTERSTITIAL"

id_dimension_value

Dimension value for the ID of this directory site. This is a read-only, auto-generated field.

active

Whether this directory site is active.

id

ID of this directory site. This is a read-only, auto-generated field.

name

Name of this directory site.