Struct google-dfareporting2::EventTagListCall

pub struct EventTagListCall<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

Retrieves a list of event tags, possibly filtered.

A builder for the list method supported by a eventTag resource. It is not used directly, but through a EventTagMethods.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.event_tags().list("profileId")
             .sort_order("consetetur")
             .sort_field("et")
             .search_string("justo")
             .add_ids("kasd")
             .add_event_tag_types("kasd")
             .enabled(true)
             .definitions_only(true)
             .campaign_id("nonumy")
             .advertiser_id("sanctus")
             .ad_id("voluptua.")
             .doit();

Methods

impl<'a, C, NC, A> EventTagListCall<'a, C, NC, A> where NC: NetworkConnector, C: BorrowMut<Client<NC>>, A: GetToken

fn doit(self) -> Result<(Response, EventTagsListResponse)>

Perform the operation you have build so far.

fn profile_id(self, new_value: &str) -> EventTagListCall<'a, C, NC, A>

Sets the profile id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

User profile ID associated with this request.

fn sort_order(self, new_value: &str) -> EventTagListCall<'a, C, NC, A>

Sets the sort order query property to the given value.

Order of sorted results, default is ASCENDING.

fn sort_field(self, new_value: &str) -> EventTagListCall<'a, C, NC, A>

Sets the sort field query property to the given value.

Field by which to sort the list.

fn search_string(self, new_value: &str) -> EventTagListCall<'a, C, NC, A>

Sets the search string query property to the given value.

Allows searching for objects by name or ID. Wildcards (*) are allowed. For example, "eventtag*2015" will return objects with names like "eventtag June 2015", "eventtag April 2015" or simply "eventtag 2015". Most of the searches also add wildcards implicitly at the start and the end of the search string. For example, a search string of "eventtag" will match objects with name "my eventtag", "eventtag 2015" or simply "eventtag".

fn add_ids(self, new_value: &str) -> EventTagListCall<'a, C, NC, A>

Append the given value to the ids query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

Select only event tags with these IDs.

fn add_event_tag_types(self, new_value: &str) -> EventTagListCall<'a, C, NC, A>

Append the given value to the event tag types query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

Select only event tags with the specified event tag types. Event tag types can be used to specify whether to use a third-party pixel, a third-party JavaScript URL, or a third-party click-through URL for either impression or click tracking.

fn enabled(self, new_value: bool) -> EventTagListCall<'a, C, NC, A>

Sets the enabled query property to the given value.

Select only enabled event tags. When definitionsOnly is set to true, only the specified advertiser or campaign's event tags' enabledByDefault field is examined. When definitionsOnly is set to false, the specified ad or specified campaign's parent advertiser's or parent campaign's event tags' enabledByDefault and status fields are examined as well.

fn definitions_only(self, new_value: bool) -> EventTagListCall<'a, C, NC, A>

Sets the definitions only query property to the given value.

Examine only the specified ad or campaign or advertiser's event tags for matching selector criteria. When set to false, the parent advertiser and parent campaign is examined as well. In addition, when set to false, the status field is examined as well along with the enabledByDefault field.

fn campaign_id(self, new_value: &str) -> EventTagListCall<'a, C, NC, A>

Sets the campaign id query property to the given value.

Select only event tags that belong to this campaign.

fn advertiser_id(self, new_value: &str) -> EventTagListCall<'a, C, NC, A>

Sets the advertiser id query property to the given value.

Select only event tags that belong to this advertiser.

fn ad_id(self, new_value: &str) -> EventTagListCall<'a, C, NC, A>

Sets the ad id query property to the given value.

Select only event tags that belong to this ad.

fn delegate(self, new_value: &'a mut Delegate) -> EventTagListCall<'a, C, NC, A>

Sets the delegate property to the given value.

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

fn param<T>(self, name: T, value: T) -> EventTagListCall<'a, C, NC, A> where T: Str

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> EventTagListCall<'a, C, NC, A> where T: Str

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead of relying on the automated algorithm which simply prefers read-only scopes over those who are not.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.