Struct google-dfareporting2::File

pub struct File {
    pub status: Option<String>,
    pub kind: Option<String>,
    pub format: Option<String>,
    pub date_range: Option<DateRange>,
    pub file_name: Option<String>,
    pub etag: Option<String>,
    pub report_id: Option<String>,
    pub urls: Option<FileUrls>,
    pub last_modified_time: Option<String>,
    pub id: Option<String>,
}

Represents a File resource. A file contains the metadata for a report run. It shows the status of the run and holds the URLs to the generated report data if the run is finished and the status is "REPORT_AVAILABLE".

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get files (response)

list files (none)

files get reports (response)

run reports (response)

Fields

status

The status of the report file.

kind

The kind of resource this is, in this case dfareporting#file.

format

The output format of the report. Only available once the file is available.

date_range

The date range for which the file has report data. The date range will always be the absolute date range for which the report is run.

file_name

The filename of the file.

etag

The eTag of this response for caching purposes.

report_id

The ID of the report this file was generated from.

urls

The URLs where the completed report file can be downloaded.

last_modified_time

The timestamp in milliseconds since epoch when this file was last modified.

id

The unique ID of this report file.