Struct google-dfareporting2::FloodlightActivityMethodsBuilder

pub struct FloodlightActivityMethodsBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on floodlightActivity resources. It is not used directly, but through the Dfareporting hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate "google-dfareporting2" as dfareporting2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use dfareporting2::Dfareporting;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Dfareporting::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `delete(...)`, `generatetag(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.floodlight_activities();

Methods

impl<'a, C, NC, A> FloodlightActivityMethodsBuilder<'a, C, NC, A>

fn get(&self, profile_id: &str, id: &str) -> FloodlightActivityGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets one floodlight activity by ID.

fn list(&self, profile_id: &str) -> FloodlightActivityListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Retrieves a list of floodlight activities, possibly filtered.

fn insert(&self, request: &FloodlightActivity, profile_id: &str) -> FloodlightActivityInsertMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Inserts a new floodlight activity.

fn delete(&self, profile_id: &str, id: &str) -> FloodlightActivityDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes an existing floodlight activity.

fn patch(&self, request: &FloodlightActivity, profile_id: &str, id: &str) -> FloodlightActivityPatchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates an existing floodlight activity. This method supports patch semantics.

fn generatetag(&self, profile_id: &str) -> FloodlightActivityGeneratetagMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Generates a tag for a floodlight activity.

fn update(&self, request: &FloodlightActivity, profile_id: &str) -> FloodlightActivityUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates an existing floodlight activity.