Struct google-dfareporting2::FloodlightConfiguration

pub struct FloodlightConfiguration {
    pub user_defined_variable_configurations: Vec<UserDefinedVariableConfiguration>,
    pub kind: Option<String>,
    pub advertiser_id: Option<String>,
    pub tag_settings: Option<TagSettings>,
    pub advertiser_id_dimension_value: Option<DimensionValue>,
    pub id: Option<String>,
    pub account_id: Option<String>,
    pub natural_search_conversion_attribution_option: Option<String>,
    pub omniture_settings: Option<OmnitureSettings>,
    pub exposure_to_conversion_enabled: Option<bool>,
    pub lookback_configuration: Option<LookbackConfiguration>,
    pub first_day_of_week: Option<String>,
    pub standard_variable_types: Vec<String>,
    pub subaccount_id: Option<String>,
    pub analytics_data_sharing_enabled: Option<bool>,
    pub id_dimension_value: Option<DimensionValue>,
    pub ssl_required: Option<bool>,
}

Contains properties of a Floodlight configuration.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get floodlight configurations (response)

update floodlight configurations (request|response)

patch floodlight configurations (request|response)

list floodlight configurations (none)

Fields

user_defined_variable_configurations

List of user defined variables enabled for this configuration.

kind

Identifies what kind of resource this is. Value: the fixed string "dfareporting#floodlightConfiguration".

advertiser_id

Advertiser ID of the parent advertiser of this floodlight configuration.

tag_settings

Configuration settings for dynamic and image floodlight tags.

advertiser_id_dimension_value

Dimension value for the ID of the advertiser. This is a read-only, auto-generated field.

id

ID of this floodlight configuration. This is a read-only, auto-generated field.

account_id

Account ID of this floodlight configuration. This is a read-only field that can be left blank.

natural_search_conversion_attribution_option

Types of attribution options for natural search conversions.

omniture_settings

Settings for DCM Omniture integration.

exposure_to_conversion_enabled

Whether the exposure-to-conversion report is enabled. This report shows detailed pathway information on up to 10 of the most recent ad exposures seen by a user before converting.

lookback_configuration

Lookback window settings for this floodlight configuration.

first_day_of_week

Day that will be counted as the first day of the week in reports. This is a required field.

standard_variable_types

List of standard variables enabled for this configuration.

Acceptable values are: - "ORD" - "NUM"

subaccount_id

Subaccount ID of this floodlight configuration. This is a read-only field that can be left blank.

analytics_data_sharing_enabled

Whether advertiser data is shared with Google Analytics.

id_dimension_value

Dimension value for the ID of this floodlight configuration. This is a read-only, auto-generated field.

ssl_required

Whether floodlight activities owned by this configuration are required to be SSL-compliant.