Struct google-dfareporting2::GeoTargeting

pub struct GeoTargeting {
    pub regions: Vec<Region>,
    pub countries: Vec<Country>,
    pub metros: Vec<Metro>,
    pub cities: Vec<City>,
    pub exclude_countries: Option<bool>,
    pub postal_codes: Vec<PostalCode>,
}

Geographical Targeting.

This type is not used in any activity, and only used as part of another schema.

Fields

regions

Regions to be targeted. For each region only dartId is required. The other fields are populated automatically when the ad is inserted or updated. If targeting a region, do not target or exclude the country of the region.

countries

Countries to be targeted or excluded from targeting, depending on the setting of the excludeCountries field. For each country only dartId is required. The other fields are populated automatically when the ad is inserted or updated. If targeting or excluding a country, do not target regions, cities, metros, or postal codes in the same country.

metros

Metros to be targeted. For each metro only dmaId is required. The other fields are populated automatically when the ad is inserted or updated. If targeting a metro, do not target or exclude the country of the metro.

cities

Cities to be targeted. For each city only dartId is required. The other fields are populated automatically when the ad is inserted or updated. If targeting a city, do not target or exclude the country of the city, and do not target the metro or region of the city.

exclude_countries

Whether or not to exclude the countries in the countries field from targeting. If false, the countries field refers to countries which will be targeted by the ad.

postal_codes

Postal codes to be targeted. For each postal code only id is required. The other fields are populated automatically when the ad is inserted or updated. If targeting a postal code, do not target or exclude the country of the postal code.