Struct google-dfareporting2::ReportMethods

pub struct ReportMethods<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on report resources. It is not used directly, but through the Dfareporting hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "google-dfareporting2" as dfareporting2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use dfareporting2::Dfareporting;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Dfareporting::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `compatible_fields_query(...)`, `delete(...)`, `files_get(...)`, `files_list(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)`, `run(...)` and `update(...)`
// to build up your call.
let rb = hub.reports();

Methods

impl<'a, C, NC, A> ReportMethods<'a, C, NC, A>

fn update(&self, request: &Report, profile_id: &str, report_id: &str) -> ReportUpdateCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates a report.

fn run(&self, profile_id: &str, report_id: &str) -> ReportRunCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Runs a report.

fn patch(&self, request: &Report, profile_id: &str, report_id: &str) -> ReportPatchCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates a report. This method supports patch semantics.

fn files_list(&self, profile_id: &str, report_id: &str) -> ReportFileListCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists files for a report.

fn insert(&self, request: &Report, profile_id: &str) -> ReportInsertCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates a report.

fn compatible_fields_query(&self, request: &Report, profile_id: &str) -> ReportCompatibleFieldQueryCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Returns the fields that are compatible to be selected in the respective sections of a report criteria, given the fields already selected in the input report and user permissions.

fn get(&self, profile_id: &str, report_id: &str) -> ReportGetCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Retrieves a report by its ID.

fn files_get(&self, profile_id: &str, report_id: &str, file_id: &str) -> ReportFileGetCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Retrieves a report file.

fn delete(&self, profile_id: &str, report_id: &str) -> ReportDeleteCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes a report by its ID.

fn list(&self, profile_id: &str) -> ReportListCall<'a, C, NC, A>

Create a builder to help you perform the following task:

Retrieves list of reports.